@@ -1,11 +1,11 @@
 name: slide_countdown_clock
 description: A countdown clock with slide up and down animation plugin for Flutter.
-version: 1.0.2
+version: 1.0.3
 author: phuc.chau <chauminhphuc.it@gmail.com>
 homepage: https://github.com/phucgaoxam/slide_countdown_clock
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
